feat: add withdraw everything to fedimint-cli
fix: fee
